backend/requirements.txt
# AniVerse Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
chromadb==0.4.22
sentence-transformers==2.3.1
pandas==2.1.4
httpx==0.26.0
python-dotenv==1.0.0
groq==0.4.2
pydantic==2.5.3
sqlalchemy==2.0.25
python-multipart==0.0.6
